Add documentation for executor.rs
